Model/requirements.txt
numpy>=1.19.0
pandas>=1.0.0
matplotlib>=3.3.0
torch>=1.7.0
torchvision>=0.8.0
scikit-learn>=0.24.0
seaborn>=0.11.0
Pillow>=8.0.0